trackday August

+++ data-plane/models/dataframe.go
@@ -6,7 +6,7 @@ type Dataframe struct {
 	SensorName string    `json:"sensorName" binding:"required"`
 	Timestamp  time.Time `json:"tmstamp,omitempty"`
 
-	SpeedKPH    float64 `json:"speedKPH,omitempty"`
+	SpeedKPH    int64   `json:"speedKPH,omitempty"`
 	Latitude    float64 `json:"latitude,omitempty"`
 	Longitude   float64 `json:"longitude,omitempty"`
 	Temperature string  `json:"temperature,omitempty"`
